+++ src/Faker.Tests/PhoneNumberTests.cs
@@ -14,5 +14,17 @@ namespace RimuTec.Faker.Tests {
          Assert.IsFalse(number.Contains("#"));
          Assert.IsFalse(string.IsNullOrWhiteSpace(number));
       }
+
+      [Test]
+      public void LandLine_HappyDays() {
+         // arrange
+
+         // act
+         var number = PhoneNumber.LandLine();
+
+         // assert
+         Assert.IsFalse(number.Contains("#"));
+         Assert.IsFalse(string.IsNullOrWhiteSpace(number));
+      }
    }
 }